# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KubeGuard Dashboard API
  version: 1.0.0
  description: >
    Backend-for-frontend for the KubeGuard dashboard. Multi-tenant, multi-cluster
    Kubernetes posture: findings ("dents"), attack paths, and compliance breach,
    tracked over time. Cluster access is read-only. Compliance figures always carry
    their `assessed` denominator and an indicative-mapping disclaimer. Secret values
    are never returned — evidence carries key names only.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/clusters:
    get:
      operationId: listClusters
      summary: List clusters in the caller's tenant.
      responses:
        "200":
          description: clusters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterList" }
    post:
      operationId: registerCluster
      summary: Register a cluster and its scannable source at runtime (admin only).
      description: >
        Adds a cluster to the caller's tenant and registers its scannable source
        (an offline manifest path or "live[:context]"). Requires the cluster.write
        permission (admin). Idempotent on the store; a rejected/malformed source is
        a 400 and leaves no cluster behind. Returns 501 when dynamic registration
        is not enabled on the server.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterClusterRequest" }
      responses:
        "201":
          description: registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cluster" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "501": { $ref: "#/components/responses/Error" }
  /v1/clusters/{id}:
    delete:
      operationId: deleteCluster
      summary: Deregister a cluster and drop its scans/history (admin only).
      description: >
        Removes a cluster from the caller's tenant along with its scans and history,
        and deregisters its scannable source so subsequent scans 404. Requires the
        cluster.write permission (admin). Tenant-scoped.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501": { $ref: "#/components/responses/Error" }
  /v1/scans:
    get:
      operationId: listScans
      parameters:
        - { name: cluster, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: scans
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanList" }
    post:
      operationId: triggerScan
      summary: Trigger an on-demand scan (analyst+).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TriggerScanRequest" }
      responses:
        "202":
          description: accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/findings:
    get:
      operationId: listFindings
      summary: Server-side filtered/sorted/paginated findings.
      parameters:
        - { name: cluster, in: query, schema: { type: string } }
        - { name: severity, in: query, schema: { type: string, description: "comma-separated: critical,high,medium,low,info" } }
        - { name: category, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, default: "severity", description: "severity|id|category" } }
        - { name: order, in: query, schema: { type: string, default: "desc", enum: [asc, desc] } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: findings page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FindingPage" }
  /v1/posture:
    get:
      operationId: getPosture
      parameters:
        - { name: cluster, in: query, schema: { type: string } }
      responses:
        "200":
          description: posture + per-framework compliance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostureResponse" }
  /v1/attack-paths:
    get:
      operationId: listAttackPaths
      parameters:
        - { name: cluster, in: query, schema: { type: string } }
      responses:
        "200":
          description: attack paths
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttackPathList" }
  /v1/history:
    get:
      operationId: getHistory
      parameters:
        - { name: cluster, in: query, schema: { type: string } }
      responses:
        "200":
          description: history snapshots (oldest first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistoryList" }
...
  /v1/audit:
    get:
      operationId: listAudit
      summary: Append-only privileged-action audit log (admin only, tenant-scoped).
      description: >
        Returns the tenant's audit entries (allowed and denied privileged actions).
        Requires the admin role; analyst/viewer receive 403.
      responses:
        "200":
          description: audit entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditList" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/report:
    get:
      operationId: getReport
      summary: Export the latest report for a scope as SARIF, CSV, or PDF.
      description: >
        Exports the most recent scan for the given cluster (or the merged fleet view
        when cluster is omitted). SARIF reuses the engine's validated 2.1.0 reporter;
        CSV is findings metadata (never secret values); PDF is a co-branded engagement
        report carrying the honest-metrics disclaimer.
      parameters:
        - { name: cluster, in: query, schema: { type: string }, description: "cluster id; omit for the merged fleet view" }
        - name: format
          in: query
          schema: { type: string, enum: [sarif, csv, pdf], default: sarif }
      responses:
        "200":
          description: "report file (served as a Content-Disposition attachment)"
          content:
            application/sarif+json:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/Error" }
        "404":
          description: no scan for this scope yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Forbidden:
      description: insufficient role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Error:
      description: bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    AuditList:
      type: object
      required: [entries]
      properties:
        entries: { type: array, items: { $ref: "#/components/schemas/AuditEntry" } }
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
    Cluster:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        environment: { type: string }
        lastScanAt: { type: string }
        totalFindings: { type: integer }
        overallPassRate: { type: number }
    ClusterList:
      type: object
      required: [clusters]
      properties:
        clusters: { type: array, items: { $ref: "#/components/schemas/Cluster" } }
...
    RegisterClusterRequest:
      type: object
      required: [id, source]
      properties:
        id: { type: string, description: "cluster id (DNS-subdomain charset)" }
        name: { type: string, description: "display name; defaults to id" }
        source: { type: string, description: "offline manifest path or \"live[:context]\"" }
    Scan:
      type: object
      required: [id, clusterId, status]
      properties:
        id: { type: string }
        clusterId: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        startedAt: { type: string }
        finishedAt: { type: string }
        totalFindings: { type: integer }
    ScanList:
      type: object
      required: [scans, total]
      properties:
        scans: { type: array, items: { $ref: "#/components/schemas/Scan" } }
        total: { type: integer }
    TriggerScanRequest:
      type: object
      required: [clusterId]
      properties:
        clusterId: { type: string }
    ResourceRef:
      type: object
      required: [kind, name]
      properties:
        kind: { type: string }
        namespace: { type: string }
        name: { type: string }
...
    Finding:
      type: object
      required: [id, title, severity, category, resource, remediation]
      properties:
        id: { type: string }
        title: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        category: { type: string }
        resource: { $ref: "#/components/schemas/ResourceRef" }
        evidence: { type: array, items: { $ref: "#/components/schemas/Evidence" } }
        remediation: { $ref: "#/components/schemas/Remediation" }
        grants: { type: array, items: { type: string } }
        refs: { type: array, items: { $ref: "#/components/schemas/ControlRef" } }
    FindingPage:
      type: object
      required: [findings, total, limit, offset]
      properties:
        findings: { type: array, items: { $ref: "#/components/schemas/Finding" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    PostureSummary:
      type: object
      required: [totalFindings, bySeverity, criticalPaths, controlsAssessed, controlsBreached, overallPassRate]
      properties:
        totalFindings: { type: integer }
        bySeverity: { type: object, additionalProperties: { type: integer } }
        criticalPaths: { type: integer }
        controlsAssessed: { type: integer }
        controlsBreached: { type: integer }
        overallPassRate: { type: number }
...
    FrameworkResult:
      type: object
      required: [framework, assessed, breached, passed, passRate, disclaimer]
      properties:
        framework: { type: string }
        version: { type: string }
        assessed: { type: integer }
        breached: { type: integer }
        passed: { type: integer }
        passRate: { type: number }
        breaches: { type: array, items: { $ref: "#/components/schemas/ControlBreach" } }
        disclaimer: { type: string }
    PostureResponse:
      type: object
      required: [posture, compliance]
      properties:
        posture: { $ref: "#/components/schemas/PostureSummary" }
        compliance: { type: array, items: { $ref: "#/components/schemas/FrameworkResult" } }
        coverage: { $ref: "#/components/schemas/CoverageBreakdown" }
        topRisks: { type: array, items: { $ref: "#/components/schemas/RiskScore" } }
...
    RiskScore:
      type: object
      required: [findingId, severity, resource, score, factors]
      description: >
        Deterministic, explainable risk score for a finding. score = sum of the
        factors' points; the factors are the full "why" (see docs/honest-metrics.md).
      properties:
        findingId: { type: string }
        title: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        resource: { $ref: "#/components/schemas/ResourceRef" }
        score: { type: integer }
        factors: { type: array, items: { $ref: "#/components/schemas/RiskFactor" } }
    CoverageBreakdown:
      type: object
      required: [discovered, assessable, skipped, rate]
      description: >
        Assessment-coverage breakdown — how much of the discovered inventory the
        engine actually assessed. rate = assessable/discovered.
      properties:
        discovered: { type: integer }
        assessable: { type: integer }
        skipped: { type: integer }
        rate: { type: number }
        skippedByKind:
          type: object
          additionalProperties: { type: integer }
...
    AttackPath:
      type: object
      required: [id, title, severity, entry, hops, summary]
      properties:
        id: { type: string }
        title: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        entry: { $ref: "#/components/schemas/ResourceRef" }
        hops: { type: array, items: { $ref: "#/components/schemas/PathHop" } }
        summary: { type: string }
    AttackPathList:
      type: object
      required: [paths]
      properties:
        paths: { type: array, items: { $ref: "#/components/schemas/AttackPath" } }
...
    HistoryList:
      type: object
      required: [snapshots]
      properties:
        snapshots: { type: array, items: { $ref: "#/components/schemas/HistorySnapshot" } }
...